Move functions to methods file
